pyproject.toml
[project]
name = "homeiot"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "influxdb-client>=1.49.0",
    "momonga>=0.2.0",
    "python-dotenv>=1.2.1",
    "ruff>=0.14.7",
]
